moving server requests into a router file
